Moved bokeh blur parameters types
